Start spotify api implementation.
Now working on the assumption that Spotify searches don't need to be cached. The variety of searches may be such that caching is not, in fact, valuable. This is especically the case with the Spotify API, as the responses are much faster than from MusicBrainz. The caching for labels should remain however, as they must be retrieved from MusicBrainz, which is much slower, and there are substantially fewer labels than releases, so caching may be more valuable.