Implement alternative search using a local copy of the MusicBrainz database.
The old method of search was as follows:
User input -> Search release from MusicBrainz (1s) -> Send data

Concurrently we had to start downloading the relevant image from coverartarchive.org, which
was also very slow.

The new method of search is as follows:
User input -> Search on spotify (300ms) -> Locate release and label in local db (~8ms) -> Send data
                                                                                       -> Lookup profile pictures (?) -> Send picture URLs
This has several benefits:

* The cover art from Spotify is much faster to download.
* We can do our own re-ranking of label search by number of releases,
  producing better results.
* The album results from Spotify are generally of higher quality.

Perhaps MusicBrainz search could be implemented as a fallback when Spotify returns
no results. We also will fail if the result is not present in the local database.
This method of search requires the release to be present both on Spotify and in the local MB
database, which may be problematic.